\name{mBvs-package}
\alias{mBvs-package}
\alias{mBvs}
\docType{package}
\title{
Multivariate Bayesian variable selection method exploiting dependence among outcomes
}
\description{
   The package provides functions to perform Bayesian variable selection for data with continuous multivariate responses and multiple covariates.
}
\details{
	The package includes the following function:
\tabular{ll}{
\code{mvnBvs} \tab Bayesian variable selection for data with continuous multivariate responses  \cr
}
\tabular{ll}{
Package: \tab mBvs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Kyu Ha Lee, Mahlet G. Tadesse, Brent A. Coull \cr
Maintainer: Kyu Ha Lee <klee@hsph.harvard.edu>
}
\references{
Lee, K. H., Tadesse, M. G., Baccarelli, A. A., Schwartz J., and Coull, B. A. (2015), 
Multivariate Bayesian Variable Selection Exploiting Dependence Structure Among Outcomes: 
Application to Air Pollution Effects on DNA Methylation, \emph{submitted}. \cr
}

\keyword{ package }
%\seealso{
%Optional links to other man pages, e.g.
%}
%\examples{
%## empty for now
%}
