\name{simData_cont}
\alias{simData_cont}
\docType{data}
\title{
A simulated data set containing multivariate normal responses and continuous covariates
}
\description{
A simulated data set containing multivariate normal responses and continuous covariates
}
\usage{data("simData_cont")}
\format{
  a list of two data frame objects. Components include, 
  \describe{
    \item{\code{Y}}{a data frame for 10 multivariate normal responses from 100 observations: \code{Y.1}-\code{Y.10}}
    \item{\code{X}}{a data frame for 2 continuous covariates from 100 observations: \code{cov.1}-\code{cov.2}}    
  }
}
%\details{
%title
%}
%\source{
%title
%}
%\references{
%title
%}
\examples{
data(simData_cont)
}
\keyword{datasets}
\keyword{multivariate continuous data}
