\name{mBvs-package}
\alias{mBvs-package}
\alias{mBvs}
\docType{package}
\title{
Bayesian Variable Selection Methods for Multivariate Data
}
\description{
Bayesian variable selection methods for data with multivariate responses and multiple covariates. The package contains implementations of multivariate Bayesian variable selection methods for continuous data and zero-inflated count data.
}
\details{
	The package includes the following function:
\tabular{ll}{
\code{mvnBvs} \tab Bayesian variable selection for data with multivariate continuous responses  \cr
\code{mzipBvs} \tab Bayesian variable selection for data with multivariate zero-inflated count responses  \cr
}
\tabular{ll}{
Package: \tab mBvs\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2021-2-3\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Kyu Ha Lee, Mahlet G. Tadesse, Brent A. Coull, Jacqueline R. Starr \cr
Maintainer: Kyu Ha Lee <klee15239@gmail.com>
}
\references{
Lee, K. H., Tadesse, M. G., Baccarelli, A. A., Schwartz J., and Coull, B. A. (2017), 
Multivariate Bayesian variable selection exploiting dependence structure among outcomes: 
application to air pollution effects on DNA methylation, \emph{Biometrics}, Volume 73, Issue 1, pages 232-241. \cr

Lee, K. H., Coull, B. A., Moscicki, A.-B., Paster, B. J., Starr, J. R. (2020),
Bayesian variable selection for multivariate zero-inflated models: application to microbiome count data, \emph{Biostatistics}, Volume 21, Issue 3, Pages 499-517 \cr
}

\keyword{ package }
%\seealso{
%Optional links to other man pages, e.g.
%}
%\examples{
%## empty for now
%}


